reuse PolicyConnRetryOptions for updater move retry options model to be used by both policy store and updater
